Split presentation cache and editor internals
